:art: blog 라우팅 수정
