+++ pages/technology/index.tsx
