+++ src/bomsbro-blog-core/templates/BlogPagesLayout.tsx
@@ -61,25 +61,25 @@ const BlogPagesLayout: React.FC<BlogPagesLayoutProps> = ({ children }) => {
       <nav className="w-full py-4 border-t border-b bg-gray-100">
         <div className="w-full flex-grow hidden sm:flex sm:items-center sm:w-auto">
           <div className="w-full container mx-auto flex flex-col sm:flex-row items-center justify-center text-sm font-bold uppercase mt-0 px-6 py-2">
-            <Link href="/blog/home">
+            <Link href="/">
               <a className="hover:bg-gray-400 rounded py-2 px-4 mx-2">Home</a>
             </Link>
-            <Link href="/blog/profile">
+            <Link href="profile">
               <a className="hover:bg-gray-400 rounded py-2 px-4 mx-2">
                 Profile
               </a>
             </Link>
-            <Link href="/blog/project">
+            <Link href="projects">
               <a className="hover:bg-gray-400 rounded py-2 px-4 mx-2">
-                Project
+                Projects
               </a>
             </Link>
-            <Link href="/blog/technology">
+            <Link href="technology">
               <a className="hover:bg-gray-400 rounded py-2 px-4 mx-2">
                 Technology
               </a>
             </Link>
-            <Link href="/blog/posts">
+            <Link href="posts">
               <a className="hover:bg-gray-400 rounded py-2 px-4 mx-2">posts</a>
             </Link>
           </div>
